Moved OpenSSL engine into standalone c and h files
References #1301
