+++ MANIFEST.in
@@ -8,6 +8,7 @@ include README.rst
 
 recursive-include docs *
 recursive-include src/cryptography/hazmat/primitives/src *.c *.h
+recursive-include src/cryptography/hazmat/bindings/openssl/src *.c *.h
 prune docs/_build
 recursive-include tests *.py
 recursive-exclude vectors *
