+++ src/cryptography/hazmat/bindings/openssl/src/osrandom_engine.h
@@ -1,0 +1,6 @@
+#ifdef _WIN32
+#include <Wincrypt.h>
+#else
+#include <fcntl.h>
+#include <unistd.h>
+#endif
